Fix issue in `meteo` slash command
There was an issue when a user asked the weather for a location
just after adding it. It was because there was no data about the
weather for this location yet. Now, the data from the first weather
request, which is used to check the existence of the specified city
in the VisualCrossing database, is reused as first sample and can
be sent to the user when the latter call the `meteo` slash command